backend/requirements.txt
Django==4.1.13
djangorestframework==3.12.4
django-filter==23.2
djoser==2.1.0
drf-extra-fields==3.7.0
Pillow==10.4.0
shortuuid==1.0.13